finished Input component
